--- micro-common/src/main/java/com/mabubu0203/sudoku/domain/package-info.java
@@ -1,1 +1,0 @@
-package com.mabubu0203.sudoku.domain;
