+++ micro-interfaces/build.gradle
@@ -1,6 +1,8 @@
 dependencies {
 
     // SpringBoot projects
+    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
+    implementation("org.springframework.boot:spring-boot-starter-data-rest")
     implementation("org.springframework.boot:spring-boot-starter-json")
     implementation("org.springframework.boot:spring-boot-starter-validation")
