+++ micro-interfaces/src/main/java/com/mabubu0203/sudoku/InterfacesCore.java
@@ -1,0 +1,19 @@
+package com.mabubu0203.sudoku;
+
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * <br>
+ *
+ * @author uratamanabu
+ * @version 1.0
+ * @since 1.0
+ */
+@Configuration
+@ComponentScan
+@EntityScan(basePackages = {"com.mabubu0203.sudoku.interfaces.domain"})
+public class InterfacesCore {
+
+}
