+++ micro-interfaces/src/main/java/com/mabubu0203/sudoku/interfaces/domain/package-info.java
@@ -1,0 +1,1 @@
+package com.mabubu0203.sudoku.interfaces.domain;
